Add required flag to registerReceiver call in AppRestrictionController
Android T adds support to allow a runtime receiver to be registered as
not exported, but to ensure apps can take advantage of this, calls to
registerReceiver must specify a flag indicating whether the receiver
should be exported for apps targeting T+ that are registering for
non-system broadcasts. This commit adds the RECEIVER_NOT_EXPORTED
flag to the receiver in AppRestrictionController since this broadcast
is sent with the app's own UID in a PendingIntent from a notification.

Bug: 161145287
Test: Build
Change-Id: I3400e9fe7e742dd9ced0d4553cbbc4e042028f47