expose HAL keyphrase trigger data to EventPayload
The SoundTrigger HAL can deliver additional context about the kephrase
that was triggered along with the trigger audio. This change moves to
expose the information which was previously suppressed by the Android
framework.

Data exposed:
- Raw data field

Previously raw data was only exposed when the data was of a specific
audio format. This change allows the raw data to be retrieved regardless
of format.

Test: atest AlwaysOnHotwordDetectorEventPayloadTest
CTS-Coverage-Bug: 215375531
Bug: 215066299
Change-Id: Ie75773853d3b338e092dfbf3bd6ccef61a1420c9