Rename VibrationThread fields/methods for clarity.
This change is purely renaming and clarification: no logic changes aside from a very minor
refactor at "StepQueue.initializeForEffect" (was "offer") to keep the steps internal.

These are groundwork for greater clarity in a larger refactor to follow.

 Key changes:
 - Rename "callback" items to have more context: whether they are actual callbacks or just
   calls into the manager. One is renamed to "hooks" instead because most methods aren't really callbacks.
 - Rename variables and methods around the handling of the "vibrator complete" callbacks to have their context.
 - Rename "CompleteStep" to EffectComplete step to confuse less with vibrator-complete callbacks.
 - Rename StepQueue methods to more accurately reflect their current narrow usage.
 - Change the word "anticipate" to better describe the case, such as resume/activate.

I also had to fix a @GuardedBy annotation to reference mLock properly from the nested object to avoid a lint error.

Test: presubmit
Bug: 193792066
Change-Id: I552572ce60264b56252b5e3dfd546ee602bc1490